requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Azure SDKs
azure-cosmos==4.8.0
azure-identity==1.15.0
openai==1.12.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Text Processing
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
black==23.12.1
